+++ packages/antd-management-fast-design-playground/CHANGELOG.md
@@ -1,5 +1,17 @@
 # antd-management-fast-design-playground
 
+## 1.0.78
+
+### Patch Changes
+
+- update
+
+## 1.0.77
+
+### Patch Changes
+
+- update
+
 ## 1.0.76
 
 ### Patch Changes