+++ packages/example-antd-management-fast-framework/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 1.8.145
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-design-playground@1.0.78
+
+## 1.8.144
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-design-playground@1.0.77
+
 ## 1.8.143
 
 ### Patch Changes